feat: implement movie API service functions for fetching movie details and popular movies, along with validation and type definitions
