+++ src/core/movie/services/api/types/movie-data-list.ts
@@ -1,0 +1,6 @@
+import { Movie } from "@core/movie/types/movie";
+
+export interface MoviesDataList {
+    movies: Movie[];
+    totalPages: number;
+  }
