+++ src/core/movie/services/api/types/movie-response.ts
@@ -1,0 +1,8 @@
+import { Movie } from "@core/movie/types/movie";
+
+export interface MovieResponse {
+    results: Movie[];
+    page: number;
+    total_pages: number;
+    total_results: number;
+  }
