+++ src/pages/movie-detail/movie-detail.tsx
@@ -1,5 +1,5 @@
 import { useParams, useNavigate, useLocation } from 'react-router-dom';
-import { getMovie } from '@movie/services/api/movie-api'
+import { getMovie } from '@core/movie/services/api/get-movie';
 import { useService } from '@app/lib/use-service';
 import { ErrorComponent } from '@core/error/components/error-component';
 import { MovieDetailSkeleton } from '@app/lib/components/skeleton-movie-detail/skeleton-movie-detail';
